ML Regression Model on User Data
Evaluate and diagnose User as 'likely to graduate from the debt management program' or not, in order to provide data analytics capabilities to the vast amount of data that our Nonprofit has, but hasn't been able to make full use of. By classifying these Users, we provide an intelligent approach to diagnose overarching reasons of which Users are most likely to default or succeed.